Move pageTitle to separate component
