+++ test/public/components/index.js
@@ -12,7 +12,9 @@
  */
 
 const itemsCounterSuite = require('./itemsCounter.test');
+const pageTitleSuite = require('./pageTitle.test');
 
 module.exports = () => {
     describe('Items counter', itemsCounterSuite);
+    describe('Page title', pageTitleSuite);
 };
