+++ test/public/utils/dataProcessing.test.js
@@ -19,9 +19,7 @@ const { extractPeriodName,
         isDetectorField,
         shouldDisplayDetectorField,
         rowDisplayStyle,
-        capitalizeFirstLetter,
-        getReadableFileSizeString,
-        pageTitle } = req('../../../app/public/utils/dataProcessing/dataProcessingUtils');
+        getReadableFileSizeString } = req('../../../app/public/utils/dataProcessing/dataProcessingUtils');
 
 module.exports = () => {
     describe('Extract period name', () => {
@@ -147,51 +145,6 @@ module.exports = () => {
 
         it('should parse GB correctly' , () => {
             assert.equal(getReadableFileSizeString(fileSizeGB), '3.5 GB' );
-        });
-    });
-
-    describe('Page title', () => {
-        const pageNames = {
-            periods: 'periods',
-            dataPasses: 'dataPasses',
-            mc: 'mc',
-            anchoredPerMC: 'anchoredPerMC',
-            anchoragePerDatapass: 'anchoragePerDatapass',
-            runsPerPeriod: 'runsPerPeriod',
-            runsPerDataPass: 'runsPerDataPass',
-            flags: 'flags',
-        };
-
-        it('should return Periods page name' , () => {
-            assert.equal(pageTitle(pageNames.periods, pageNames), 'Periods');
-        });
-
-        it('should return Data passes page name' , () => {
-            assert.equal(pageTitle(pageNames.dataPasses, pageNames), 'Data passes per period');
-        });
-
-        it('should return Monte Carlo page name' , () => {
-            assert.equal(pageTitle(pageNames.mc, pageNames), 'Monte Carlo');
-        });
-
-        it('should return Anchored per MC page name' , () => {
-            assert.equal(pageTitle(pageNames.anchoredPerMC, pageNames), 'Anchored per MC');
-        });
-
-        it('should return Anchorage per data pass page name' , () => {
-            assert.equal(pageTitle(pageNames.anchoragePerDatapass, pageNames), 'Anchorage per data pass');
-        });
-
-        it('should return Runs per period page name' , () => {
-            assert.equal(pageTitle(pageNames.runsPerPeriod, pageNames), 'Runs per period');
-        });
-
-        it('should return Runs per data pass page name' , () => {
-            assert.equal(pageTitle(pageNames.runsPerDataPass, pageNames), 'Runs per data pass');
-        });
-
-        it('should return Quality flags page name' , () => {
-            assert.equal(pageTitle(pageNames.flags, pageNames), 'Quality flags');
         });
     });
 };
